fetch/api/headers/header-values.html

<!doctype html>
<meta charset=utf8>
<meta name=timeout content=long>
<title>Header value test</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<div id=log></div>
<script>
for(let i = 0; i < 0x100; i++) {
  let fail = false
  if(i === 0 || i === 0x0A || i === 0x0D) {
    fail = true
  }

  let val = "x" + String.fromCharCode(i) + "x"
  async_test((t) => {
    let xhr = new XMLHttpRequest()
    xhr.open("POST", "../resources/inspect-headers.py?headers=value-test")
    if(fail) {
      assert_throws("SyntaxError", () => xhr.setRequestHeader("value-test", val))
      t.done()
    } else {
      xhr.setRequestHeader("value-test", val)
      xhr.onload = t.step_func_done(() => {
        assert_equals(xhr.getResponseHeader("x-request-value-test"), val)
      })
      xhr.send()
    }
  }, "XMLHttpRequest with value " + encodeURI(val))

  promise_test((t) => {
    if(fail) {
      return promise_rejects(t, new TypeError(), fetch("about:blank", { headers: {"value-test": val} }))
    } else {
      return fetch("../resources/inspect-headers.py?headers=value-test", { headers: {"value-test": val} }).then((res) => {
        assert_equals(res.headers.get("x-request-value-test"), val)
      })
    }
  }, "fetch() with value " + encodeURI(val))
}
</script>
